detection-engine/requirements.txt
### Core Dependencies
redis==5.0.1
opensearch-py==2.4.0
pyyaml==6.0.1
python-dateutil==2.8.2

### Parsing and Processing
geoip2==4.7.0
dnspython==2.4.2
requests==2.31.0

### Data Validation
jsonschema==4.20.0
cerberus==1.3.5

### Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

### Utilities
python-dotenv==1.0.0
click==8.1.7

### Testing (dev)
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1

### Code Quality (dev)
black==23.12.0
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1
